% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_plot.R
\name{enrichMito}
\alias{enrichMito}
\title{Mitochondrial Enrichment Analysis of a gene list.}
\usage{
enrichMito(genes, database)
}
\arguments{
\item{genes}{a vector of gene ENSEMBL id.}

\item{database}{character string saying the database to use for the analysis.
Either one of "MitoCarta", "Reactome", "GO-CC" and "GO-BP".}
}
\value{
enrichment analysis for the mitochondrial gene sets.
}
\description{
Given a vector of genes, this function will return the 
enrichment analysis for the mitochondrial gene sets after FDR control. For 
the Reactome, GO-CC and GO-BP databases it returns also the enrichment 
results for the corresponding original pathways.
}
