% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_plot.R
\name{gseaMito}
\alias{gseaMito}
\title{Mitochondrial GSEA of a gene list.}
\usage{
gseaMito(genes, database)
}
\arguments{
\item{genes}{order ranked gene vector named by ENSEMBL id.}

\item{database}{character string saying the database to use for the analysis.
Either one of "MitoCarta", "Reactome", "GO-CC" and "GO-BP".}
}
\value{
GSEA results for the mitochondrial gene sets.
}
\description{
Gene set enrichment analysis for the mitochondrial gene sets. 
For the Reactome, GO-CC and GO-BP databases it returns also the GSEA 
results for the corresponding original pathways.
}
\examples{
data(ovse)

}
