% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_plot.R
\name{mitoAnalysisSections}
\alias{mitoAnalysisSections}
\title{Perform ssGSEA analysis by sections.}
\usage{
mitoAnalysisSections(dataset = NULL, database = "MitoCarta")
}
\arguments{
\item{dataset}{Gene expression data which can be given either as a SummarizedExperiment or ExpressionSet object, or as a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{database}{One of MitoCarta, Reactome, GO-CC and GO-BP. Default is MitoCarta.}
}
\description{
Given a dataset, it returns the ssGSEA score for each section and each sample.
}
