% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MitoGenesDB}
\alias{MitoGenesDB}
\title{Mitochondrial genes}
\format{
An object of class \code{data.frame} with 2996 rows and 3 columns.
}
\usage{
data(MitoGenesDB)
}
\description{
Here are listed all the final mitochondrial genes in ENSEMBL id, the 
corresponding SYMBOL id and the database from where they were collected.
}
\keyword{datasets}
