% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitology-package.R
\docType{package}
\name{mitology-package}
\alias{mitology}
\alias{mitology-package}
\title{mitology: Study of mitochondrial activity from RNA-seq data}
\description{
mitology allows to study the mitochondrial activity throught high-throughput RNA-seq data. It is based on a collection of genes whose proteins localize in to the mitochondria. From these, mitology provides a reorganization of the pathways related to mitochondria activity from Reactome and Gene Ontology. Further a ready-to-use implementation of MitoCarta3.0 pathways is included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CaluraLab/mitology}
  \item Report bugs at \url{https://github.com/CaluraLab/mitology/issues}
}

}
\author{
\strong{Maintainer}: Stefania Pirrotta \email{stefania.pirrotta@unipd.it} (\href{https://orcid.org/0009-0004-0030-217X}{ORCID})

Authors:
\itemize{
  \item Enrica Calura \email{enrica.calura@unipd.it} (\href{https://orcid.org/0000-0001-8463-2432}{ORCID}) [funder]
}

}
\keyword{internal}
