% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_plot.R
\name{mitoPlotLeaves}
\alias{mitoPlotLeaves}
\title{Plot a circular heatmap of leaf ssGSEA scores on the dendrogram.}
\usage{
mitoPlotLeaves(
  leafResults = NULL,
  database = "MitoCarta",
  samples = NULL,
  outerNames = "sections"
)
}
\arguments{
\item{leafResults}{Result of the mitoAnalysisLeaves function.}

\item{database}{One of MitoCarta, Reactome, GO-CC and GO-BP. Default is MitoCarta.}

\item{samples}{A vector with the names of samples to be plotted. If NULL all samples are plotted.}

\item{outerNames}{Whether to plot section names or leaves names. Default is sections.}
}
\description{
Plot a circular heatmap of leaf ssGSEA scores on the dendrogram.
}
