% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_plot.R
\name{mitoPlotSections}
\alias{mitoPlotSections}
\title{Plot a circular heatmap of section ssGSEA scores on the dendrogram.}
\usage{
mitoPlotSections(sectionResults = NULL, database = "MitoCarta", samples = NULL)
}
\arguments{
\item{sectionResults}{Result of the mitoAnalysisSections function.}

\item{database}{One of MitoCarta, Reactome, GO-CC and GO-BP. Default is MitoCarta.}

\item{samples}{A vector with the names of samples to be plotted. If NULL all samples are plotted.}
}
\description{
Plot a circular heatmap of section ssGSEA scores on the dendrogram.
}
