% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_plot.R
\name{mitoPlotLeaves}
\alias{mitoPlotLeaves}
\title{Circular heatmap of mitochondrial gene set ssGSEA scores.}
\usage{
mitoPlotLeaves(
  leafResults,
  database = "MitoCarta",
  sections = FALSE,
  samples = NULL,
  labelNames = "sections",
  ...
)
}
\arguments{
\item{leafResults}{output of the enrichment.}

\item{database}{character string saying the database used for the analysis.
Either one of "MitoCarta", "Reactome", "GO-CC" and "GO-BP".}

\item{sections}{logical. Either to keep the aggregated gene set categories 
or the specific gene sets. Default is FALSE.}

\item{samples}{character vector with the names of samples to be plotted.
Otherwise all samples are plotted.}

\item{labelNames}{character string that says to plot the names of "sections"
or "leaves".}

\item{...}{other arguments passed on to the \code{\link[ggtree]{gheatmap}}
function.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Given the enrichment output, it returns a circular heatmap of 
the ssGSEA scores for each specific mitochondrial gene set (leaf of the 
database tree) or gene set group (section of the database tree).
}
\examples{
MClist <- getGeneSets()

}
